--- src/api.rs
@@ -1,4 +1,0 @@
-use super::admindb;
-use super::admindb::AdminDB;
-
-pub mod handlers;
