--- src/api/handlers.rs
@@ -1,31 +1,0 @@
-use serde::{Deserialize, Serialize};
-use std::sync::Arc;
-
-use super::{admindb, AdminDB};
-pub use error::Error;
-
-pub type Result<T> = std::result::Result<T, Error>;
-
-#[derive(Deserialize, Serialize)]
-pub struct EmailPassword {
-    email: String,
-    password: String,
-}
-
-/// Authenticates email and password combination
-pub async fn authenticate_email_password(
-    _admindb: Arc<AdminDB>,
-    _cred: EmailPassword,
-) -> Result<bool> {
-    Ok(false)
-}
-
-pub mod error {
-    /// Handler errors
-    #[derive(thiserror::Error, Debug)]
-    pub enum Error {
-        /// Database errors
-        #[error(transparent)]
-        DB(#[from] super::admindb::Error),
-    }
-}
